fix: licensed-gate compile clean in all four configs
First full compile of the staged runtime against the installed game
(2.0.211.51): five missing using directives, four short-circuit paths
leaving out-params/locals unassigned (fail-closed defaults, refusals
preserved), engine CoreVersion is System.Version not string (adapted our
side via ToString, matching the engine's own stringification), two inner-
local shadowing renames, two int->short casts. Dev-harness compile route
receipt now exists: devModes baseline/compatibility rc 0/0, 23 shards.

TafTests 10,179/0. PortableTests 2,187/0.

+++ Simulation/City/KingdomPorters.02a.ConstructionInputSourceRendering.cs
@@ -1,5 +1,6 @@
 using XRL;
 using XRL.World;
+using XRL.World.Parts;
 
 namespace ThousandAndFirst.Simulation.City
 {
@@ -23,8 +24,8 @@ namespace ThousandAndFirst.Simulation.City
 			if (verdict == KingdomBindingVerdict.Mint)
 			{
 				minted = true;
-				GameObject created = Mint(system, zone, tripId, row.DeliverySourceX,
-					row.DeliverySourceY, row.OriginCode, now);
+				GameObject created = Mint(system, zone, tripId, (short)row.DeliverySourceX,
+					(short)row.DeliverySourceY, row.OriginCode, now);
 				if (!GameObject.Validate(created)) return false;
 				r_KingdomPorter part = created.RequirePart<r_KingdomPorter>();
 				part.JobId = tripId;
